- Finished Product details design changes v2
